+++ src/main/java/goorming/iCurriculum/curriculum/CurriculumTable.java
@@ -1,4 +1,4 @@
-package goorming.iCurriculum.curriculumtable;
+package goorming.iCurriculum.curriculum;
 
 import goorming.iCurriculum.common.BaseEntity;
 import goorming.iCurriculum.department.Department;
