+++ src/main/java/goorming/iCurriculum/essentialcourse/EssentialCourse.java
@@ -37,5 +37,4 @@ public class EssentialCourse extends BaseEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "department_id")
     private Department department;
-
 }
